+++ sdk/securityinsight/arm-securityinsight/src/securityInsights.ts
@@ -46,6 +46,7 @@ import {
   ThreatIntelligenceIndicatorMetricsImpl,
   WatchlistsImpl,
   WatchlistItemsImpl,
+  WorkspaceManagerGroupsImpl,
   DataConnectorsImpl,
   DataConnectorsCheckRequirementsOperationsImpl,
   OperationsImpl
@@ -82,6 +83,7 @@ import {
   ThreatIntelligenceIndicatorMetrics,
   Watchlists,
   WatchlistItems,
+  WorkspaceManagerGroups,
   DataConnectors,
   DataConnectorsCheckRequirementsOperations,
   Operations
@@ -174,7 +176,7 @@ export class SecurityInsights extends coreClient.ServiceClient {
 
     // Assigning values to Constant parameters
     this.$host = options.$host || "https://management.azure.com";
-    this.apiVersion = options.apiVersion || "2022-09-01-preview";
+    this.apiVersion = options.apiVersion || "2023-03-01-preview";
     this.alertRules = new AlertRulesImpl(this);
     this.actions = new ActionsImpl(this);
     this.alertRuleTemplates = new AlertRuleTemplatesImpl(this);
@@ -214,6 +216,7 @@ export class SecurityInsights extends coreClient.ServiceClient {
     );
     this.watchlists = new WatchlistsImpl(this);
     this.watchlistItems = new WatchlistItemsImpl(this);
+    this.workspaceManagerGroups = new WorkspaceManagerGroupsImpl(this);
     this.dataConnectors = new DataConnectorsImpl(this);
     this.dataConnectorsCheckRequirementsOperations = new DataConnectorsCheckRequirementsOperationsImpl(
       this
@@ -281,6 +284,7 @@ export class SecurityInsights extends coreClient.ServiceClient {
   threatIntelligenceIndicatorMetrics: ThreatIntelligenceIndicatorMetrics;
   watchlists: Watchlists;
   watchlistItems: WatchlistItems;
+  workspaceManagerGroups: WorkspaceManagerGroups;
   dataConnectors: DataConnectors;
   dataConnectorsCheckRequirementsOperations: DataConnectorsCheckRequirementsOperations;
   operations: Operations;